### 2020-02-01 #2051 @whedon
[PRE REVIEW]: PyEscape
**Submitting author:** @sirsharpest (<a href="http://orcid.org/0000-0002-4572-5828">Nathan Hughes</a>)
**Repository:** <a href="https://github.com/SirSharpest/NarrowEscapeSimulator" target ="_blank">https://github.com/SirSharpest/NarrowEscapeSimulator</a>
**Version:** 1.0
