### 2020-02-01 #2051 @Kevin-Mattheus-Moerman
@SirSharpest I think your paper is not compiling because you used a ` symbol in the YAML code at the top for the title. Can you replace:
```YAML
title: PyEscape: A narrow escape problem simulator package for python'
```
with
```YAML
title: PyEscape: A narrow escape problem simulator package for python
```
Once you made the change you can call `@whedon generate pdf` here to reattempt the creation of the the pdf.
