### 2020-02-01 #2051 @Kevin-Mattheus-Moerman
@openjournals/dev am I right to say it is the ` symbol? Or are those allowed in YAML?
