### 2020-02-01 #2051 @xuanxu
> openjournals/dev am I right to say it is the ` symbol? Or are those allowed in YAML?

The error is that the parser considers `PyEscape:` a YAML key, out of place, so it errors.
To keep the title, quoting the whole string will make it:
```YAML
title: "PyEscape: A narrow escape problem simulator package for python"
```
