### 2020-02-01 #2051 @arfon
@SirSharpest - could you please truncate your bibtex file to only include the entries that this paper actually cites?
