+++ library.bib
@@ -1,42 +1,0 @@
-@misc{NumpyUniqueNumPy,
-  title = {Numpy.Unique \textemdash{} {{NumPy}} v1.16 {{Manual}}},
-  file = {/Users/hughesn/Zotero/storage/L6ST64W3/numpy.unique.html},
-  howpublished = {https://docs.scipy.org/doc/numpy/reference/generated/numpy.unique.html},
-  note = {00000}
-}
-
-@article{holcmanEscapeSmallOpening2004,
-  title = {Escape {{Through}} a {{Small Opening}}: {{Receptor Trafficking}} in a {{Synaptic Membrane}}},
-  shorttitle = {Escape {{Through}} a {{Small Opening}}},
-  author = {Holcman, D. and Schuss, Z.},
-  year = {2004},
-  month = dec,
-  volume = {117},
-  pages = {975--1014},
-  issn = {1572-9613},
-  doi = {10.1007/s10955-004-5712-8},
-  abstract = {We model the motion of a receptor on the membrane surface of a synapse as free Brownian motion in a planar domain with intermittent trappings in and escapes out of corrals with narrow openings. We compute the mean confinement time of the Brownian particle in the asymptotic limit of a narrow opening and calculate the probability to exit through a given small opening, when the boundary contains more than one. Using this approach, it is possible to describe the Brownian motion of a random particle in an environment containing domains with small openings by a coarse grained diffusion process. We use the results to estimate the confinement time as a function of the parameters and also the time it takes for a diffusing receptor to be anchored at its final destination on the postsynaptic membrane, after it is inserted in the membrane. This approach provides a framework for the theoretical study of receptor trafficking on membranes. This process underlies synaptic plasticity, which relates to learning and memory. In particular, it is believed that the memory state in the brain is stored primarily in the pattern of synaptic weight values, which are controlled by neuronal activity. At a molecular level, the synaptic weight is determined by the number and properties of protein channels (receptors) on the synapse. The synaptic receptors are trafficked in and out of synapses by a diffusion process. Following their synthesis in the endoplasmic reticulum, receptors are trafficked to their postsynaptic sites on dendrites and axons. In this model the receptors are first inserted into the extrasynaptic plasma membrane and then random walk in and out of corrals through narrow openings on their way to their final destination.},
-  journal = {Journal of Statistical Physics},
-  language = {en},
-  number = {5}
-}
-
-@article{schussNarrowEscapeProblem2007,
-  title = {The Narrow Escape Problem for Diffusion in Cellular Microdomains},
-  author = {Schuss, Z. and Singer, A. and Holcman, D.},
-  year = {2007},
-  month = oct,
-  volume = {104},
-  pages = {16098--16103},
-  issn = {0027-8424, 1091-6490},
-  doi = {10.1073/pnas.0706599104},
[…]
