### 2020-02-04 #2051 @drvinceknight
@markgalassi and @pdebuyl would you be able to review this software paper for JOSS? It's a package for a particle type of Brownian particle simulation which I believe is within both of your expertise. If you've not done it before, reviewing for JOSS is slightly different to a traditional journal, you can find information about the process here https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'd be delighted to answer any questions you may have.
