### 2020-02-04 #2051 @pdebuyl
Hi @drvinceknight

I will review the paper. I already worked for JOSS for the paper on the rowan quaternion library.
