### 2020-02-04 #2051 @markgalassi
Dear @drvinceknight I will be happy to review the paper.  I started another one once but got badly ill for a protracted period, but that was long ago so I am ready.
