### 2020-02-05 #2072 @whedon
[REVIEW]: PyEscape
**Submitting author:** @sirsharpest (<a href="http://orcid.org/0000-0002-4572-5828">Nathan Hughes</a>)
**Repository:** <a href="https://github.com/SirSharpest/NarrowEscapeSimulator" target ="_blank">https://github.com/SirSharpest/NarrowEscapeSimulator</a>
**Version:** 1.0
