### 2020-02-10 872ed9b
Corrected unit vector value in travel function
