### 2020-02-13 #2072 @pdebuyl
Hi @SirSharpest ,

I had a first look at the program. I believe that documentation needs to be improved. I have
written a first review below. I suggest that you take a look at existing scientific codes on GitHub, preferably with a JOSS paper, to have a good impression of what is expected. The purpose of the review is to eventually tick all the boxes, so I can provide more information if needed.

## Functionality

### Installation: Does installation proceed as outlined in the documentation?

`pip install` does not work because of permissions. Instead of using `sudo pip install`
(which users could easily find by googling), I suggest to propose `pip install --user`
instead.

Functionality: Have the functional claims of the software been confirmed?

I ran the Example notebook with success. I would like to have a least a few benchmark cases.

## Documentation

A statement of need: Do the authors clearly state what problems the software is designed to
solve and who the target audience is?

The problem is well stated. The target audience is not.

Installation instructions: Is there a clearly-stated list of dependencies? Ideally these
should be handled with an automated package management solution.

A requirements.txt file is missing. This is the most standard way to state dependencies for
Python projects.

Example usage: Do the authors include examples of how to use the software (ideally to solve
real-world analysis problems).

There is one example notebook. I suggest to improve the example by adding a textbook-type
example, with known numerical output.

Functionality documentation: Is the core functionality of the software documented to a
satisfactory level (e.g., API method documentation)?

No. There are no docstring and no module-wide documentation page. The purpose of the
routines is not documented explicitly, only the example, the name of the functions, and the
source code provide that information.

Automated tests: Are there automated tests or manual steps described so that the
functionality of the software can be verified?

There is limited testing. The test checks against an upper limit. An output time of zero
would pass the test. There is no theoretic estimate provided for the test. Playing with the
test case repeatedly, I had a 1% failure rate.

Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute
to the software 2) Report issues or problems with the software 3) Seek support

There is a contribution section in the readme. It only applies to new features and not to
bug reports, other issues, or support requesets.
