### 2020-02-13 6ffd32a
Added some docstrings and more tests
