+++ narrow_escape/escape_drawing.py
@@ -6,6 +6,10 @@ from .escape_utility import sphere_vol_to_r, cube_vol_to_r
 
 
 def draw_sphere(v, ax):
+    """Draws a sphere on an axis
+
+    Sphere volume and axis to draw on need to be specified
+    """
     r = sphere_vol_to_r(v)
     u = np.linspace(0, 2 * np.pi, 100)
     v = np.linspace(0, np.pi, 100)
@@ -17,6 +21,10 @@ def draw_sphere(v, ax):
 
 
 def draw_cube(v, ax):
+    """Draws a cube on axis
+
+    Cube volume and axis to draw on need to be specified
+    """
     r = cube_vol_to_r(v)
     r = [-r/2, r/2]
     for s, e in combinations(np.array(list(product(r, r, r))), 2):
