+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup
 
 setup(name='narrow_escape',
-      version='0.2',
+      version='0.3',
       description='Library used for simulating narrow escape problems',
       url='https://github.com/SirSharpest/narrow_escape',
       author='Nathan Hughes',

[…]
