### 2020-02-14 #2072 @SirSharpest
I have just pushed a series of commits which I feel address most of the concerns:

# Functionality

- A `--user` flag had been added
- Installation uses setuptools to install and setup the application and the requirements. A requirements.txt would be redundant as setuptools handles packages
- The examples notebook, and the readme.md have been updated to reflect software functionality

# Documentation
- A sentence was added to the readme.md to give use-cases. However, we see this software as having use-cases that we currently do not foresee as it is can be used in various branches of research
- Documentation relating to installation has been revised
- Our examples notebook now uses a real-world example from published work
- Docstrings have been added to all functions
- Automated testing of major modules and features are implemented.
- Tests with a failure rate because of time have been given an appropriate flag with a pytest decorator
- Added info for contributing and contacting with requests

# Other

I am happy to take on board any other issues that reviewers may have and will address them promptly also. Thank you again for your valuable input.
