### 2020-02-20 #2072 @pdebuyl
@SirSharpest the doi for the JOSS paper should already be known. @drvinceknight is it ok to cite a JOSS paper before its acceptance?
