### 2020-02-20 #2072 @SirSharpest
I don't think it would be appropriate to cite before acceptance, [the DOI](https://doi.org/10.21105/joss.02072) currently is inactive it seems.
