### 2020-02-21 ac36fc9
Simplified travel eq
