### 2020-02-21 #2072 @markgalassi
I just got back from travel and dove in to the paper.  I am afraid that it is clearly not a candidate for anything yet.  The simplest of things (following the procedure in README.md) fails:
`pores = fibonacci_spheres(p, v)   `

gives:

`NameError: name 'p' is not defined`

Clearly some variable p should be set, but for a minimal example which is the first thing shown this is not given!

Please have the author double-check that the README.md works and then we can get back to work on evaluating this software.
