### 2020-02-21 #2072 @markgalassi
Let me also add a comment on the idea of notebooks as documentation: I know they are much in vogue, but there are people (like me) who insist on reproducible procedures for running python code as batch.  This means that I am not in the habit of loading notebooks, and if I am presented with documentation as a notebook I want a command line invocation which will load all that.  So pointing to a directory is not enough: please give a command to load that notebook up from a standard GNU/Linux system.
