+++ narrow_escape/escape_plan.py
@@ -6,8 +6,8 @@ from .escape_detection import in_sphere, in_cube, passthrough_pore, passthrough_
 def travel(delta, pa):
     """Find a new position for a particle
 
-    Takes a delta, movement size and a particle of N dimensions
-    returns a new array of similar dimensions to pa.
+    Takes a delta, movement size and a particle of N dimensions returns a new
+    array of similar dimensions to pa.
 
     """
     p = pa.copy()
@@ -23,13 +23,15 @@ def escape_with_path(r, delta, dt, shape, max_steps,
                      pore_locs, pore_size, check_func, cur_pos):
     """Provides the full path of a particle as it escapes from a container
 
-    Takes a radius of container, delta step size, dt difference of time,
-    a shape (cube of sphere), a maximum number of steps to allow, location of escape pore(s),
-    size of escape pore(s), a checking function for collision and
-    the current position of the particle
+    Takes a radius of container, delta step size, dt difference of time, a
+    shape (cube of sphere), a maximum number of steps to allow, location of
+    escape pore(s), size of escape pore(s), a checking function for collision
+    and the current position of the particle
 
-    returns a multi-dimensional array of a particle at each position as it escapes the container
-    will return if max steps is reached or when escape is detected
+    returns a multi-dimensional array of a particle at each position as it
+    escapes the container will return if max steps is reached or when escape is
+    detected
+
     """
     path = np.zeros((max_steps, 3))
     path[0] = cur_pos
@@ -51,12 +53,15 @@ def escape_with_path(r, delta, dt, shape, max_steps,
 def escape(D, vol, pore_size, pore_locs,
            dt=None, seed=None, shape='sphere',
            max_steps=(int(1e7)), with_path=False, random_start=False, flat=False):
-    """Wrapper function that can be called by a user - used to optimise code shared between escape methods
+    """Wrapper function that can be called by a user - used to optimise code
+    shared between escape methods
 
-    Takes a diffusion coefficent, volume of container, escape pore size, escape pore(s) location.
-    Optional arguments of step-size and a numpy random seed can be given, as well as specifying maximum number of
-    steps that a walk can take. Switches to return the full path (or just the time required), to start the particle
-    at the container centroid or random placement, and to detect a flat exit area or an exit volume.
+    Takes a diffusion coefficent, volume of container, escape pore size, escape
+    pore(s) location. Optional arguments of step-size and a numpy random seed
+    can be given, as well as specifying maximum number of steps that a walk can
+    take. Switches to return the full path (or just the time required), to
+    start the particle at the container centroid or random placement, and to
+    detect a flat exit area or an exit volume.
 
     """
     if dt is None:
@@ -94,12 +99,13 @@ def escape_flat(r, delta, dt, shape, max_steps,
                 pore_locs, pore_size, check_func, cur_pos):
     """Simulates a escape of a a particle, through a flat pore on surface of container
 
-    Takes a radius of container, delta step size, dt difference of time,
-    a shape (cube of sphere), a maximum number of steps to allow, location of escape pore(s),
-    size of escape pore(s), a checking function for collision and
-    the current position of the particle
+    Takes a radius of container, delta step size, dt difference of time, a
+    shape (cube of sphere), a maximum number of steps to allow, location of
+    escape pore(s), size of escape pore(s), a checking function for collision
+    and the current position of the particle
 
     returns a number of steps taken to escape
+
     """
     check_func = in_sphere if shape == 'sphere' else in_cube
     steps = 0
@@ -119,12 +125,13 @@ def escape_quick(r, delta, dt, shape, max_steps,
                  pore_locs, pore_size, check_func, cur_pos):
     """Simulates escape without tracking position through container
 
[…]
