### 2020-02-24 #2072 @SirSharpest
Thanks for your comments.

I have added the required lines to the read me.

Re Notebooks: For the case of being documentation and providing an example of how to use the software, then the integrated github viewer should be enough. This is presented as a method without results (beyond solving a known problem), so I wouldn't expect running it to be a complete necessity. Saying that however, most users will want to run notebooks differently and depending on your current setup you should be able to run `jupyter-notebook <notebook name>` from your command-line.

Though, if you're wanting that command-line experience then running
 ```
narrow_escape -D 400 -v 1 -a 0.1 -p 1 -N 1000 -dt 2e-9

# "narrow_escape --help" can give arguments info
```
from bash (providing you installed the library) will give the same results, without plots as in the example notebook.
