### 2020-02-24 36e07c1
Added numpy import
