+++ readme.md
@@ -66,7 +66,7 @@ To get a more accurate value, multiple simulations are required e.g. :
 ``` python
 from narrow_escape.escape_plan import escape
 from narrow_escape.escape_points import fibonacci_spheres, points_on_cube_surface
-
+import numpy as np
 
 D = 400
 v = 1
@@ -88,7 +88,7 @@ To extend this further, we much wish to multi-process to speed up simulations:
 ``` python
 from narrow_escape.escape_plan import escape
 from narrow_escape.escape_points import fibonacci_spheres, points_on_cube_surface
-
+import numpy as np
 
 D = 400
 v = 1
