### 2020-02-26 b3bd5bc
removed unused time import and variable
