+++ Testing/test_escape.py
@@ -1,5 +1,4 @@
 import pytest
-from time import time
 import numpy as np
 from narrow_escape.escape_plan import escape
 from narrow_escape.escape_points import fibonacci_spheres, points_on_cube_surface
@@ -8,7 +7,6 @@ from narrow_escape.escape_utility import sphere_vol_to_r, calculate_delta
 
 @pytest.mark.timeout(60)
 def test_escape():
-    t1 = time()
     D = 400
     v = 1
     a = 0.1
