### 2020-02-26 ce26cc9
add instructions to install and run pytest
