### 2020-02-26 #2072 @pdebuyl
Thank you for the update. I ticked many checks for the review already. Some items remain to
improve, especially given the focus of JOSS on best practices (including testing).

1. The comment "base install of python (not recommended)" is superfluous.

2. The tests are not well defined. For instance, the test for the sphere placement routine
   does not check the distance of the pore but that it falls below a much larger upper bound.

3. The tests use floating-point equality, which is not a reliable way to perform numerical
   tests. The most convenient solutions are either NumPy's testing routines (that include
   "closeness" tests) or pytest's approx feature.

4. The escape tests now does not include a bound. The earlier version was better but an
   estimate for this time must be given. Only "0" would not pass the test now.

5. scipy is not used in the code, it should be removed from the install_requires.

6. There are unused imports and unused variables in the test files, they should be removed.

7. The duplicate "%notebook" directive does not work as intented in the notebook example
   (the directive can only be used once per notebook).

8. The optional argument max_steps does not limit the range of searches that exit the
   container. Picking a dt that is too large results in infinite loops.

9. If max_steps is reached, the escape routine returns max_steps*dt instead of an obviously
   erroneous values (0 or np.nan).

I filed pull requests for 7 and part of 6.
