+++ Testing/test_points.py
@@ -1,8 +1,5 @@
 from narrow_escape.escape_points import fibonacci_spheres, points_on_cube_surface
 from narrow_escape.escape_detection import in_sphere, in_cube
-import pytest
-import numpy as np
-
 
 def test_fibo_spheres():
     assert in_sphere(fibonacci_spheres(1, 1)[0], 1)
