+++ Testing/test_escape.py
@@ -1,8 +1,7 @@
-import pytest
-import numpy as np
 from narrow_escape.escape_plan import escape
 from narrow_escape.escape_points import fibonacci_spheres, points_on_cube_surface
-from narrow_escape.escape_utility import sphere_vol_to_r, calculate_delta
+from narrow_escape.escape_utility import sphere_vol_to_r
+import pytest
 
 
 @pytest.mark.timeout(60)
@@ -25,3 +24,13 @@ def test_escape_cube():
     pores = points_on_cube_surface(1, r)
     t = escape(D, v, a, pores, shape='cube')
     assert t
+
+
+def test_timeout():
+    D = 400
+    v = 1
+    a = 0.1
+    r = sphere_vol_to_r(v)
+    pores = points_on_cube_surface(1, r)
+    t = escape(D, v, a, pores, shape='cube', max_steps=(1))
+    assert int(t) == 0
