### 2020-03-09 #2072 @markgalassi
Further testing: thanks @SirSharpest for addressing the issues I reported previously.  Smaller ones now:
1. Since you document how to run *before* you document installation, you should slip in a "after installing (according to the instructions below)" here and there.
2. Because of the python community's sad failure in updating the world to python3, you might want to mention that on many systems they might have to run "pip3 install ."
3. The second test python program fails with a:
```
$ python3 t2.py
Traceback (most recent call last):
  File "t2.py", line 9, in <module>
    pores = fibonacci_spheres(p, v)
NameError: name 'p' is not defined
```
So p and dt are not defined.  This is the usual problem with the notebook people trying to write reproducible and deliverable software and documentation :-)
So the author should definitely test all those examples.

Otherwise my earlier problems seem to have been addressed and I am continuing with the checklist.
