+++ readme.md
@@ -76,6 +76,8 @@ import numpy as np
 D = 400
 v = 1
 a = 0.1
+p = 1
+dt = 1e-6 # dt approaching 0 will take longer but give more accurate results
 
 pores = fibonacci_spheres(p, v)
 
