### 2020-03-10 #2072 @SirSharpest
Again, thank you for the comments and suggestions. Particularly towards testing and the errors in the examples, I was running them sequentially and so variables were carried over.

I've made updates to the readme to remove the unnecessary text also.
I'm reluctant to mention pip3 commands, simply because I state twice that this is a python 3 library and adding pip3 would cause errors for some setups. For example:

```
zsh> which pip
/Users/nathan/anaconda3/bin/pip
zsh> which pip3
/usr/local/bin/pip3
zsh> conda activate myenv; which pip
/Users/nathan/anaconda/envs/bin/pip
```

I've also updated several tests to use the numpy floating point closeness functions.
Removed scipy from the setup.py.
Cleaned up some unused imports in testing
Added a test for max_steps variable being hit
Reaching max_steps now returns a zero value (I am working on writing some warnings to be raised also)
Some tests still need to be addressed (in progress)

Pull requests for minor fixes have been accepted.
