### 2020-03-11 #2072 @markgalassi
I have checked most of the boxes in the paper checklist.  Here is a small suggested diff (below) so you don't assume that users are steeped in markdown.  I would suggest that the "statement of need" in the readme.md file could do with a bit more, and that the paper put a math citation at the moment of stating "The mathematical models provided are simple and robust @iWouldCiteMySourceHere".  I also have not yet checked the "state of the field" checkbox.  In the paper you discuss the Schuss and Holcman papers, but no reference to existing software.  You say yours is "novel" but you do not say that there is no "narrow escape" softare.

And another nit: when I did a "git diff" after running your notebook procedure I saw a diff on a date in the notebook.  Remember: notebooks are not reproducible and they are not entirely human source, so it is flawed to add them to a version control repo.  I would not hold up the paper on this, but I would recommend that you put reproducibility and best VC practices front-and-center: commit a .py file, then provide a simple one-liner to load that into a notebook for notebook types.

Other those simple suggestions I think you meet the criteria and I would quickly finish this off.

And if I might add a personal note: your project is interesting!  It made me want to read up on narrow escape and take an interest in the topic.

```
diff --git a/paper.md b/paper.md
index c78409e..bd9404c 100644
--- a/paper.md
+++ b/paper.md
@@ -1,3 +1,9 @@
+---
+# process with:
+# pandoc --bibliography paper.bib -o paper.pdf paper.md
+# pandoc --standalone --bibliography paper.bib -o paper.html paper.md
+---
+
 ---
 title: "PyEscape: A narrow escape problem simulator package for python"
 tags:
```
