### 2020-03-11 cf2eff2
Updated text, added two more references
