+++ paper.bib
@@ -40,3 +40,38 @@
   number = {41},
   pmid = {17901203}
 }
+
+@article{codlingRandomWalkModels2008,
+  title = {Random Walk Models in Biology},
+  author = {Codling, Edward A and Plank, Michael J and Benhamou, Simon},
+  year = {2008},
+  month = aug,
+  volume = {5},
+  pages = {813--834},
+  doi = {10.1098/rsif.2008.0014},
+  abstract = {Mathematical modelling of the movement of animals, micro-organisms and cells is of great relevance in the fields of biology, ecology and medicine. Movement models can take many different forms, but the most widely used are based on the extensions of simple random walk processes. In this review paper, our aim is twofold: to introduce the mathematics behind random walks in a straightforward manner and to explain how such models can be used to aid our understanding of biological processes. We introduce the mathematical theory behind the simple random walk and explain how this relates to Brownian motion and diffusive processes in general. We demonstrate how these simple models can be extended to include drift and waiting times or be used to calculate first passage times. We discuss biased random walks and show how hyperbolic models can be used to generate correlated random walks. We cover two main applications of the random walk model. Firstly, we review models and results relating to the movement, dispersal and population redistribution of animals and micro-organisms. This includes direct calculation of mean squared displacement, mean dispersal distance, tortuosity measures, as well as possible limitations of these model approaches. Secondly, oriented movement and chemotaxis models are reviewed. General hyperbolic models based on the linear transport equation are introduced and we show how a reinforced random walk can be used to model movement where the individual changes its environment. We discuss the applications of these models in the context of cell migration leading to blood vessel growth (angiogenesis). Finally, we discuss how the various random walk models and approaches are related and the connections that underpin many of the key processes involved.},
+  file = {/Users/hughesn/Google Drive/PHD/Zotfile/C/Codling et al. - 2008 - Random walk models in biology.pdf;/Users/hughesn/Zotero/storage/X2QPKKHK/rsif.2008.html},
+  journal = {Journal of The Royal Society Interface},
+  note = {01012},
+  number = {25}
+}
+
+
+@article{kayeFastSolverNarrow2020,
+  title = {A Fast Solver for the Narrow Capture and Narrow Escape Problems in the Sphere},
+  author = {Kaye, Jason and Greengard, Leslie},
+  year = {2020},
+  month = jan,
+  volume = {5},
+  pages = {100047},
+  issn = {2590-0552},
+  doi = {10.1016/j.jcpx.2019.100047},
[…]
