### 2020-03-11 c85a9f2
Modified readme with statement of need
