+++ readme.md
@@ -10,6 +10,20 @@ Examples are provided in the `notebooks` folder.
 This software can be used by anyone wanting to simulate how long a particle under Brownian motion will take to escape from a container with a number of escape pores on it's surface. This was developed with a cellular biology context, however usage in chemistry, physics and animal sciences could easily be imagined. This software is presented as easily modifiable.
 
 
+## Statement of need 
+
+Many tutorials, examples and solutions to random walks are widely available.
+Solutions with boundary conditions are rare, but still given. For the more
+complex, such as in three dimensions and in specific shapes this is difficult
+problem with no standard solution. 
+
+We provide this library to give a simple protocol for estimating narrow escape
+problems. By using stochastic simulations results can be quickly found. 
+
+In particular, this library is useful for working with specifically, or
+randomly, placed exit pores with varying sizes and with non-standard shapes ( we
+provide a cube example). 
+
 ## A note on running
 
 This software can be run on a standard PC, we have tested for both OSX and Linux. However, running large numbers of simulations is best done on high-performance-computing equiment. This package is optimised for larger number of CPUs. Runnining on older/slower hardware may require additional time.
