### 2020-03-11 #2072 @SirSharpest
I do see your points on issues with notebooks - I will move away from these in future (currently investigating your suggestion to make them non issues with tracking).

And thank you for your interest, we were surprised when we found how little narrow escape problems have been applied to problems outside of physics. We hope this library will help with that by making it more accessible. As per your suggestion I've added a recent paper giving a mathematical solution for a similar problem, but which lacks a programmatically useable one.
