### 2020-03-15 #2072 @markgalassi
I just reviewed again and see that "A Statement of Need" is now in the readme.md file.  In the paper you also have more of a reference to existing software efforts.  I have checked off the missing boxes.
