### 2020-03-16 #2072 @markgalassi
I think I might be done and I'm happy to recommend the software+paper as it is now. I also find that my fellow reviewer was a bit more subtle in some of their comments so I would encourage the author to incorporate those comments, but I would not require it for readiness. Editor: do I have to take any specific steps beyond stating this here?
