### 2020-03-18 #2072 @pdebuyl
Hi @SirSharpest (sorry for the delay. I still work but organization is different. Belgium ordered confinement, all nonessential displacements are forbidden) , there remain issues among the ones I raised.

> 2. The tests are not well defined. For instance, the test for the sphere placement routine
>    does not check the distance of the pore but that it falls below a much larger upper bound.

   In the test below, the sphere is of volume 1 (thus radius of about 0.62) but the test `in_sphere` is for a radius of 1, so the test is too tolerant.
```
  def test_fibo_spheres():
    assert in_sphere(fibonacci_spheres(1, 1)[0], 1)
```

A variation such as `in_sphere(fibonacci_spheres(1, 1)[0], R+eps)` where R is "(3/(4*pi))**(1/3)" and eps a suitably small value (order accuracy of floating point variables) would be much more useful. A similar remark applies to test_detection.py

> 3. The tests use floating-point equality, which is not a reliable way to perform numerical
>    tests. The most convenient solutions are either NumPy's testing routines (that include
>    "closeness" tests) or pytest's approx feature.

Here the change is good, except that requiring "decimal=2" for `test_cube_vol_to_r` and `test_sphere_vol_to_r` feels overly inaccurate.

> 4. The escape tests now does not include a bound. The earlier version was better but an
>    estimate for this time must be given. Only "0" would not pass the test now.

This has not been addressed. `assert t` will only fail for `t=0` and this is worse than the previous generous upper bound.

> 8. The optional argument max_steps does not limit the range of searches that exit the
>    container. Picking a dt that is too large results in infinite loops.

This has not been addressed. In more detail: the increase of max_steps occurs in the outer loop:

https://github.com/SirSharpest/NarrowEscapeSimulator/blob/master/narrow_escape/escape_plan.py#L41

The while loop just below will go on indefinitely if one does not find a suitable `new_pos`, for instance if dt is too large.
