### 2020-03-20 1a0e96d
Added more refined points placement testing
