+++ Testing/test_points.py
@@ -1,9 +1,21 @@
 from narrow_escape.escape_points import fibonacci_spheres, points_on_cube_surface
 from narrow_escape.escape_detection import in_sphere, in_cube
-
-def test_fibo_spheres():
-    assert in_sphere(fibonacci_spheres(1, 1)[0], 1)
+from narrow_escape.escape_utility import sphere_vol_to_r
 
 
-def test_ponts_on_cube():
-    assert in_cube(points_on_cube_surface(1, 1)[0], 1.1)
+V = 1
+P = 1
+eps = .001 * V 
+
+def test_fibo_spheres_inside():
+
+    assert in_sphere(fibonacci_spheres(P, V)[0], sphere_vol_to_r(V+eps))
+
+def test_fibo_spheres_outside():
+    assert in_sphere(fibonacci_spheres(P, V)[0], sphere_vol_to_r(V-eps)) == False
+
+def test_ponts_on_cube_inside():
+    assert in_cube(points_on_cube_surface(P, V)[0], V+eps)
+
+def test_ponts_on_cube_outside():
+    assert in_cube(points_on_cube_surface(P, V)[0], V-eps) == False
