### 2020-03-20 d93dee0
Added all testing to 5 decimal places
