### 2020-03-20 72095d7
Fixed infinite loop potential bug
