### 2020-03-20 #2072 @SirSharpest
Thank you @pdebuyl, I understand that for many these are difficult times, I hope yourself and other reviewers and editors are under no pressure with regards to JOSS / this submission as it is not in any way time-dependant and your own wellbeings are much more vital.

I'll address your feedback in order:

1. I've just added 2 new tests, and modified the others that you've pointed out. I now test for tighter bounds to ensure that functions are working correctly.

2. Upped the numpy testing to 5 decimal places

3.  Regarding timeouts and escape times. I am unsure of how to best handle this. Due to the nature of this being stochastic I'm wary of putting concrete bounds. I've attempted to address this concern by implementing another test which runs for N simulations and ensures that the given result is within a reasonable expected range. This test also has a timeout attached on it, should this test fail then it is a reasonable indication that something is misbehaving.

4.  Thank you for pointing out the inner while look problem, I now see. I've implemented a correction in the latest commit. As a side note, I recognise this has increased issues with code repetition. An issue I was hoping to iron out, and will refactor in the future.

4b. I've added a test with a large dt that breaks similarly to max_steps when required.
