### 2020-03-20 #2072 @pdebuyl
Hi @SirSharpest thank you for the update. I understand that stochastic simulations are difficult to test!

I am satisfied will the reply to all my comments. I have a few minor remaining concerns, none preventing acceptance but I prefer to mention them.

1. I filed a PR to add the import of Axes3D from mpl_toolkits.mplot3d to enable the 3d plot in the notebook.
2. NumPy should be cited in the paper. "We used NumPy [citation] for the random number generation, for testing, and for storing the numerical data in arrays." With the citation recommended at https://scipy.org/citing.html#numpy pinging @drvinceknight for the policy about this.
3. The repository name is NarrowEscape, the pull request here and the paper title are PyEscape, the actual library is narrow_escape. I suggest a bit of cleaning up there. If you plan to put the code on PyPI at some point, this will confuse the users.

Apart from the citation issue (the only unticked box), I approve the paper for publication and leave the other issues to your choice.
