### 2020-03-20 #2072 @pdebuyl
+nitpick: in the example notebook, the timestep is 2e-9 whereas calculate_opt_dt returns 4.2e-6, this is rather inconsistent. If 4.2e-6 is too small, calculate_opt_dt should return a smaller value.
