### 2020-03-20 ac50065
Refactored to make better naming, updated paper to reflect libraries used
