+++ paper.bib
@@ -5,6 +5,58 @@
   note = {00000}
 }
 
+@conference{Kluyver:2016aa,
+	Author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing},
+	Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+	Editor = {F. Loizides and B. Schmidt},
+	Organization = {IOS Press},
+	Pages = {87 - 90},
+	Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
+	Year = {2016}}
+
+@article{hunterMatplotlib2DGraphics2007a,
+  title = {Matplotlib: {{A 2D Graphics Environment}}},
+  shorttitle = {Matplotlib},
+  author = {Hunter, John D.},
+  year = {2007},
+  month = may,
+  volume = {9},
+  pages = {90--95},
+  issn = {1558-366X},
+  doi = {10.1109/MCSE.2007.55},
+  abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting,and publication-quality image generation across user interfaces and operating systems},
+  file = {/Users/nathan/Zotero/storage/D5AEJ5WM/4160265.html},
+  journal = {Computing in Science Engineering},
+  keywords = {2D graphics package,application development,computer graphics,Computer languages,Equations,Graphical user interfaces,Graphics,Image generation,interactive scripting,Interpolation,mathematics computing,Matplotlib,object-oriented programming,operating system,Operating systems,Packaging,Programming profession,publication-quality image generation,Python,scientific programming,scripting languages,software packages,user interface,User interfaces},
+  note = {09180 
+Conference Name: Computing in Science Engineering},
+  number = {3}
+}
+
+@article{vanderwaltNumPyArrayStructure2011,
+  title = {The {{NumPy Array}}: {{A Structure}} for {{Efficient Numerical Computation}}},
+  shorttitle = {The {{NumPy Array}}},
+  author = {{van der Walt}, Stefan and Colbert, S. Chris and Varoquaux, Gael},
+  year = {2011},
+  month = mar,
+  volume = {13},
+  pages = {22--30},
+  issn = {1558-366X},
+  doi = {10.1109/MCSE.2011.37},
+  abstract = {In the Python world, NumPy arrays are the standard representation for numerical data and enable efficient implementation of numerical computations in a high-level language. As this effort shows, NumPy performance can be improved through three techniques: vectorizing calculations, avoiding copying data in memory, and minimizing operation counts.},
+  file = {/Users/nathan/Google Drive/PHD/Zotfile/V/van der Walt et al. - 2011 - The NumPy Array A Structure for Efficient Numeric.pdf;/Users/nathan/Zotero/storage/4RUAD4C9/5725236.html},
+  journal = {Computing in Science Engineering},
+  keywords = {Arrays,Computational efficiency,data structures,Finite element methods,high level language,high level languages,mathematics computing,numerical analysis,Numerical analysis,numerical computation,numerical computations,numerical data,NumPy,numpy array,Performance evaluation,programming libraries,Python,Python programming language,Resource management,scientific programming,Vector quantization},
+  note = {04954 
+Conference Name: Computing in Science Engineering},
+  number = {2}
+}
+
+
+
+
+
+
 @article{holcmanEscapeSmallOpening2004,
[…]
