### 2020-03-20 d543a92
Removed reference to speed
