### 2020-03-20 #2072 @SirSharpest
Apologies for making Whedon work extra hard, I've made a couple of final changes for now:

1. I've refactored the naming of the library to suit the paper title better, as per the suggestion, it was overdue on my part
2. I added a section to the paper citing Numpy, Matplotlib and Jupyter as they are key external libraries used
3. Calculations for optimal dt has been improved by 2 orders of magnitude, code will run slower but be more accurate, ultimately.

Again, thank you all for the feedback and help with this.
