### 2020-03-20 #2072 @pdebuyl
Thanks for the update. It's all ok for me but you probably want to undelete the code though. `git add PyEscape` and `git commit` should do :-)

@drvinceknight I recommend for publication.
