### 2020-03-24 #2072 @drvinceknight
@SirSharpest everything looks good to me.

Could you make a tagged release and archive it (on Zenodo). When you've done that, report the version number and archive DOI here please. Please make sure the archive deposit has the correct metadata (title and author list) to match the paper.

If you could also confirm the version number, it's currently `1.0` if that should be updated as you've modified things throughout the review please let me know.
