### 2020-03-24 #2072 @SirSharpest
Hi @drvinceknight,

Version will be at 1.0 as to me this published version should be marked as that.
The DOI is 10.5281/zenodo.3725946 (it's coming up as not working as I write this, but as far as I am aware DOI takes a while to appear online)

Do let me know if I need to do anything more or if I've made any errors.

Thanks!
