### 2020-03-31 1b24233
minor fixes to paper
