### 2020-03-31 #2072 @danielskatz
Additionally, I've suggested some changes to the paper in https://github.com/SirSharpest/NarrowEscapeSimulator/pull/6
