### 2020-03-31 #2072 @danielskatz
And I've suggested some changes to the references in https://github.com/SirSharpest/NarrowEscapeSimulator/pull/7
